requirements.txt
fastapi[standard]
pydantic
python-decouple
langchain
langchain-ollama
discord.py
slack-bolt
docker
paramiko